Move WithRouteTagMiddleware test into sub test package
