+++ instrumentation/net/http/otelhttp/test/handler_test.go
@@ -468,3 +468,29 @@ func TestSpanStatus(t *testing.T) {
 		})
 	}
 }
+
+func TestWithRouteTagMiddleware(t *testing.T) {
+	recorder := tracetest.NewSpanRecorder()
+	tp := sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(recorder))
+	ctx, _ := tp.Tracer(t.Name()).Start(context.Background(), "test-route-tag-attribute")
+
+	expectedRouteTag := "a route"
+
+	nextCalled := false
+	mw := otelhttp.WithRouteTagMiddleware(expectedRouteTag)(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
+		nextCalled = true
+	}))
+
+	r, err := http.NewRequestWithContext(ctx, http.MethodGet, "http://localhost/", nil)
+	require.NoError(t, err)
+
+	rr := httptest.NewRecorder()
+	mw.ServeHTTP(rr, r)
+
+	assert.True(t, nextCalled)
+
+	spans := recorder.Started()
+	if assert.Len(t, spans, 1) {
+		assert.Contains(t, spans[0].Attributes(), semconv.HTTPRoute(expectedRouteTag))
+	}
+}
